dashboard/requirements.txt
streamlit==1.38.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
db-dtypes==1.3.0
pyarrow==17.0.0
